refactor(utils): move extract_archive into its own module
The next commit deletes this repo's download transport in favour of
deepcell-auth, but ``extract_archive`` stays: deepcell-auth's extractor
calls ``extractall`` without vetting members, so delegating it would drop
the zip-slip / tar-traversal / symlink rejection and the member-count and
size bounds. Move it (and the archive limits it reads) out of ``_auth``
ahead of that deletion, and split its tests out of ``tests/test_auth.py``.

Pure move: no behavior change.
